playback: split sinks per-class, move out of player/
Extract playback subsystem from src/player/ to src/playback/. Split
playback_sinks.cpp into one file per sink under playback/sinks/
(cef_thread_sinks, browser_sink, idle_inhibit_sink, theme_color_sink,
media_session_sink, mpv_action_sink). sinks.h becomes umbrella header.
Update includes and CMake source lists.
