+++ CMakeLists.txt
@@ -129,9 +129,14 @@ if(APPLE)
         src/player/media_session.cpp
         src/player/macos/media_session_macos.mm
         src/player/media_session_thread.cpp
-        src/player/playback_state_machine.cpp
-        src/player/playback_coordinator.cpp
-        src/player/playback_sinks.cpp
+        src/playback/state_machine.cpp
+        src/playback/coordinator.cpp
+        src/playback/sinks/cef_thread_sinks.cpp
+        src/playback/sinks/browser_sink.cpp
+        src/playback/sinks/idle_inhibit_sink.cpp
+        src/playback/sinks/theme_color_sink.cpp
+        src/playback/sinks/media_session_sink.cpp
+        src/playback/sinks/mpv_action_sink.cpp
     )
 
     # LetsMove - prompts user to move app to /Applications (clears quarantine)
@@ -160,9 +165,14 @@ elseif(WIN32)
         src/player/media_session.cpp
         src/player/windows/media_session_windows.cpp
         src/player/media_session_thread.cpp
-        src/player/playback_state_machine.cpp
-        src/player/playback_coordinator.cpp
-        src/player/playback_sinks.cpp
+        src/playback/state_machine.cpp
+        src/playback/coordinator.cpp
+        src/playback/sinks/cef_thread_sinks.cpp
+        src/playback/sinks/browser_sink.cpp
+        src/playback/sinks/idle_inhibit_sink.cpp
+        src/playback/sinks/theme_color_sink.cpp
+        src/playback/sinks/media_session_sink.cpp
+        src/playback/sinks/mpv_action_sink.cpp
         ${CMAKE_BINARY_DIR}/iconres.rc
     )
     set(PLATFORM_LIBRARIES
@@ -339,9 +349,14 @@ else()
         src/player/mpris/media_session_mpris.cpp
         src/player/mpris/mpris_projection.cpp
         src/player/media_session_thread.cpp
-        src/player/playback_state_machine.cpp
-        src/player/playback_coordinator.cpp
-        src/player/playback_sinks.cpp
+        src/playback/state_machine.cpp
+        src/playback/coordinator.cpp
+        src/playback/sinks/cef_thread_sinks.cpp
+        src/playback/sinks/browser_sink.cpp
+        src/playback/sinks/idle_inhibit_sink.cpp
+        src/playback/sinks/theme_color_sink.cpp
+        src/playback/sinks/media_session_sink.cpp
+        src/playback/sinks/mpv_action_sink.cpp
         # Wayland
         src/platform/wayland.cpp
         src/clipboard/wayland.cpp
