+++ src/playback/coordinator.cpp
@@ -1,4 +1,4 @@
-#include "playback_coordinator.h"
+#include "coordinator.h"
 
 #include <utility>
