+++ src/playback/sinks.h
@@ -1,0 +1,9 @@
+#pragma once
+
+// Umbrella header — individual sinks live in sinks/.
+#include "sinks/cef_thread_sinks.h"
+#include "sinks/browser_sink.h"
+#include "sinks/idle_inhibit_sink.h"
+#include "sinks/theme_color_sink.h"
+#include "sinks/media_session_sink.h"
+#include "sinks/mpv_action_sink.h"
