+++ src/playback/state_machine.cpp
@@ -1,4 +1,4 @@
-#include "playback_state_machine.h"
+#include "state_machine.h"
 
 #include <utility>
