+++ src/playback/state_machine.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "playback_event.h"
+#include "event.h"
 
 #include <cstdint>
 #include <string>
